VSR: Fix liveness for io_depth_write overflow when replicas <= 2
Where `config.pipelining_max` exceeds `config.io_depth_write` it's
possible for a client request to be unable to acquire a write IOP if we
have maxed out our IO depth.

This can lead to deadlock for a cluster of one or two, since there is
no other way for the leader to repair the dirty op because no other
replica has it.

The fix is for `on_prepare_timeout()` to retry the prepare.

Reported-by: @ThreeFx

Fixes: #5