VOPR: Add preliminary crash injection (not yet in scope)
This is NOT YET in scope for the bounty, while we wrap up the
Recovery protocol from the 2012 VRR paper, which needs a little
more time only because we're also recovering from stable storage
using "Protocol-Aware Recovery for Consensus-Based Storage".

This work is almost done however, so soon we'll have a whole
new way of finding correctness and liveness bugs.